chore: move client struct to own file
